requirements.txt
biopython>=1.80
wheel>=0.38.2
ncbi-datasets-pylib>=15.19
prettytable
pandas
networkx
scikit-learn
dask
statsmodels
seaborn
swifter